requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
langgraph>=0.2.0
pydantic>=2.8.0
python-docx>=1.1.2
python-pptx>=0.6.23
openpyxl>=3.1.5
pypdf>=5.0.0
python-multipart>=0.0.9
